Divide ClassicApp into interface and implementation
