+++ patterns/adapter/flutter_adapter/adapter/classic_app.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import '../../../observer/app_observer/observer/app_observer.dart';
-import '../../../observer/app_observer/observer/event.dart';
 
 abstract class ClassicApp {
   AppObserver get events;
@@ -25,45 +24,4 @@ abstract class ClassicApp {
   void onPaint(Canvas canvas, Size size);
 
   void repaint();
-}
-
-class ClassicAppRepaint extends Event {}
-
-abstract class ClassicAppBase implements ClassicApp {
-  final _appEvents = AppObserver();
-
-  @override
-  AppObserver get events => _appEvents;
-
-  @override
-  void onKeyDown() {}
-
-  @override
-  void onKeyUp() {}
-
-  @override
-  void onMoseEnter() {}
-
-  @override
-  void onMouseDown() {}
-
-  @override
-  void onMouseLeave() {}
-
-  @override
-  void onMouseMove() {}
-
-  @override
-  void onMouseUp() {}
-
-  @override
-  void onWheel(double deltaX, double deltaY) {}
-
-  @override
-  void onPaint(Canvas canvas, Size size);
-
-  @override
-  void repaint() {
-    _appEvents.notify(ClassicAppRepaint());
-  }
 }
