+++ patterns/adapter/flutter_adapter/adapter/classic_app_events.dart
@@ -1,0 +1,3 @@
+import '../../../observer/app_observer/observer/event.dart';
+
+class ClassicAppRepaint extends Event {}
